+++ lib/graph_kit.dart
@@ -37,5 +37,6 @@ export 'src/graph_algorithms.dart';
 export 'src/node.dart';
 export 'src/node_type.dart';
 export 'src/pattern_query.dart';
+export 'src/pattern_query_petit.dart';
 export 'src/serialization.dart';
 export 'src/traversal.dart';